Backend/Backend101/lesson-5/pyproject.toml
[tool.poetry]
name = "my-backend-project"
version = "0.1.0"
description = "todo-list"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
numpy = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
